made response type enum package private. added static get value methid
