+++ src/main/java/com/redislabs/redisgraph/ResultSet.java
@@ -7,20 +7,6 @@ import java.util.Iterator;
  */
 public interface ResultSet extends Iterator<Record> {
 
-
-    enum ResultSetScalarTypes {
-        VALUE_UNKNOWN,
-        VALUE_NULL,
-        VALUE_STRING,
-        VALUE_INTEGER,
-        VALUE_BOOLEAN,
-        VALUE_DOUBLE,
-        VALUE_ARRAY,
-        VALUE_EDGE,
-        VALUE_NODE
-    }
-    
-
     int size();
 
     Statistics getStatistics();
