+++ src/main/java/com/redislabs/redisgraph/graph_entities/GraphEntity.java
@@ -1,8 +1,5 @@
 package com.redislabs.redisgraph.graph_entities;
 
-
-import com.redislabs.redisgraph.ResultSet.ResultSetScalarTypes;
-
 import java.util.*;
